pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "canopen"
authors = [
    {name = "Christian Sandberg", email = "christiansandberg@me.com"},
    {name = "André Colomb", email = "src@andre.colomb.de"},
    {name = "André Filipe Silva", email = "afsilva.work@gmail.com"},
]
description = "CANopen stack implementation"
readme = "README.rst"
requires-python = ">=3.8"
license = {file = "LICENSE.txt"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "python-can >= 3.0.0",
]
dynamic = ["version"]

[project.optional-dependencies]
db_export = [
    "canmatrix ~= 1.0",
]

[project.urls]
documentation = "https://canopen.readthedocs.io/en/stable/"
repository = "https://github.com/canopen-python/canopen"

[tool.setuptools]
packages = ["canopen"]

[tool.setuptools_scm]
version_file = "canopen/_version.py"

[tool.pytest.ini_options]
testpaths = [
    "test",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.mypy]
python_version = "3.8"
exclude = [
    "^examples*",
    "^test*",
    "^setup.py*",
]
disable_error_code = "annotation-unchecked"
